possible-fix: search not showing up in some devices #93

+++ repository/src/main/java/app/simple/felicity/repository/models/Audio.java
@@ -190,7 +190,7 @@ public class Audio implements Parcelable {
         uri = in.readString();
         track = in.readInt();
         author = in.readString();
-        size = in.readInt();
+        size = in.readLong();
         albumArtist = in.readString();
         year = in.readString();
         bitrate = in.readLong();
